Use Zero Literally When Cookie Expiration is Zero (#37)
* Use Zero Literally When Cookie Expiration is Zero

Otherwise add the cookie expiration to the current time. This means that
folks using the `0` in their bundle config will have a cookie that lasts
the length of the browser session. Everyone else will get a cookie that
expiers $cookieExpire seconds in the future.

Should fix #33 and address some of the concerns raised in #35 and the
commit 5479263a0aa78221d2350d52eb51d747e0ea7d33

* Add Tests to Verify the Cookie Expiration Behavior

- One test to verify that a zero value sets the cookie with no
  expiration (a session cookie).
- Another to test that an actual expiration is set to a time in the
  future.